BACK-010-FixFiltersBugs

+++ Back/BackServer/RepositoryChangers/Implementations/PropertyChangerDb.cs
@@ -11,13 +11,14 @@ namespace BackServer.RepositoryChangers.Implementations
     public class PropertyChangerDb : IPropertyChanger
     {
         private readonly TestContext _context;
+        private readonly string defaultPropertyValue = "Не указано";
 
         public PropertyChangerDb(TestContext context)
         {
             _context = context;
         }
 
-        public async Task<bool> Add(Property property)
+        public async Task<bool> AddProperty(Property property)
         {
             var propertyDb = new DbEntity.Property() {Title = property.Title};
             if (!await _context.Properties.AnyAsync(x => x.Title == property.Title))
@@ -43,8 +44,33 @@ namespace BackServer.RepositoryChangers.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
+        
+        public async Task<bool> AddPropertyValue(string propertyTitle, string[] propertyValues)
+        {
+            var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
+            if (property == null)
+            {
+                return false;
+            }
+            
+            foreach (var value in propertyValues)
+            {
+                if (!await _context.PropertyValues.AnyAsync(x =>
+                    x.Property==property && x.PropertyValue == value))
+                {
+                    var propertyValue = new DbEntity.PropertyValues()
+                        {Property = property, PropertyValue = value};
+                    await _context.PropertyValues.AddAsync(propertyValue);
+                    await _context.SaveChangesAsync();
+                }
+            }
 
-        public async Task<bool> Delete(string propertyTitle)
+
+            await _context.SaveChangesAsync();
+            return true;
+        }
+
+        public async Task<bool> DeleteProperty(string propertyTitle)
         {
             var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
             if (property == null)
@@ -53,24 +79,52 @@ namespace BackServer.RepositoryChangers.Implementations
             var propertyValuesSet = _context.PropertyValues
                 .Where(x => x.Property == property)
                 .ToHashSet();
-            var headingsThree = await _context.HeadingsThree
-                .Where(x => propertyValuesSet.Contains(x.PropertyValues))
+            foreach (var propertyValue in propertyValuesSet)
+            {
+                await DeletePropertyValue(propertyTitle, propertyValue.PropertyValue);
+            }
+            
+            var headingsOneFilters = await _context.HeadingOneFilters
+                .Where(x => x.property_id == property.Id)
                 .ToArrayAsync();
-            var productProperties = await _context.ProductProperties
-                .Where(x => propertyValuesSet.Contains(x.PropertyValues))
+            var headingsTwoFilters = await _context.HeadingTwoFilters
+                .Where(x => x.property_id == property.Id)
                 .ToArrayAsync();
 
-
-            _context.ProductProperties.RemoveRange(productProperties);
-            _context.HeadingsThree.RemoveRange(headingsThree);
+            
             _context.PropertyValues.RemoveRange(propertyValuesSet);
+            _context.HeadingOneFilters.RemoveRange(headingsOneFilters);
+            _context.HeadingTwoFilters.RemoveRange(headingsTwoFilters);
 
             _context.Properties.Remove(property);
             await _context.SaveChangesAsync();
             return true;
         }
 
-        public async Task<bool> Update(Property oldProperty, Property property)
+        public async Task<bool> DeletePropertyValue(string propertyTitle, string propertyValue)
+        {
+            var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
+            if (property == null)
+                return false;
+
+            var propertyValueDb = await _context.PropertyValues
+                .FirstOrDefaultAsync(x => x.Property == property && x.PropertyValue == propertyValue);
+            if (propertyValueDb == null)
+                return false;
+
+            var headingsThree = await _context.HeadingsThree
+                .Where(x => x.PropertyValues==propertyValueDb)
+                .ToArrayAsync();
+            await SetPropertyDefault(propertyTitle, propertyValue);
+            
+            _context.HeadingsThree.RemoveRange(headingsThree);
+
+            _context.PropertyValues.Remove(propertyValueDb);
+            await _context.SaveChangesAsync();
+            return true;
+        }
+
+        public async Task<bool> UpdateProperty(Property oldProperty, Property property)
         {
             var propertyDb = await _context.Properties.FirstOrDefaultAsync(x => x.Title == oldProperty.Title);
             if (propertyDb == null)
@@ -88,7 +142,21 @@ namespace BackServer.RepositoryChangers.Implementations
             return true;
         }
 
-        public async Task<bool> AddProductPropertyValue(string productTitle, string propertyTitle, string propertyValue, bool isPriority)
+        public async Task<bool[]> AddProductProperties(string productTitle, Entity.Property[] properties)
+        {
+            var res = new bool[properties.Length];
+            for (var i = 0; i < properties.Length; i++)
+            {
+                var property = properties[i];
+                res[i] = await AddProductPropertyValue(productTitle, property.Title, property.Values.First(),
+                    property.IsPriority);
+            }
+
+            return res;
+        }
+
+        public async Task<bool> AddProductPropertyValue(string productTitle, string propertyTitle, string propertyValue,
+            bool isPriority)
         {
             var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
             if (property == null)
@@ -97,6 +165,12 @@ namespace BackServer.RepositoryChangers.Implementations
             var product = await _context.Products.FirstOrDefaultAsync(x => x.Title == productTitle);
             if (product == null)
                 return false;
+
+            if (await _context.ProductProperties.FirstOrDefaultAsync(x =>
+                x.Product == product && x.PropertyValues.Property == property) != null)
+            {
+                return true;
+            }
 
             var propertyValues =
                 await _context.PropertyValues.FirstOrDefaultAsync(x =>
@@ -108,12 +182,12 @@ namespace BackServer.RepositoryChangers.Implementations
                     {Property = property, PropertyValue = propertyValue};
 
                 await _context.PropertyValues.AddAsync(propertyValues);
+                await _context.SaveChangesAsync();
             }
-            else if (await _context.ProductProperties.FirstOrDefaultAsync(x =>
-                x.Product == product && x.PropertyValues == propertyValues) != null)
-            {
-                return true;
-            }
+
+            propertyValues =
+                await _context.PropertyValues.FirstAsync(x =>
+                    x.Property.Id == property.Id && x.PropertyValue == propertyValue);
 
             await _context.ProductProperties.AddAsync(new ProductProperty()
             {
@@ -174,7 +248,8 @@ namespace BackServer.RepositoryChangers.Implementations
             if (propertyValues.Count == 0)
                 return false;
 
-            var productProperties = await _context.ProductProperties.Where(x => propertyValues.Contains(x.PropertyValues))
+            var productProperties = await _context.ProductProperties
+                .Where(x => propertyValues.Contains(x.PropertyValues))
                 .ToArrayAsync();
 
             _context.ProductProperties.RemoveRange(productProperties);
@@ -184,29 +259,65 @@ namespace BackServer.RepositoryChangers.Implementations
             return true;
         }
 
+        private async Task SetPropertyDefault(string propertyTitle, string propertyValue)
+        {
+            var defaultPV = await _context.PropertyValues
+                .FirstOrDefaultAsync(x => x.Property.Title == propertyTitle && x.PropertyValue == defaultPropertyValue);
+            if (defaultPV == null)
+            {
+                await AddProperty(new Property(propertyTitle, new[] {defaultPropertyValue}));
+            }
+            
+            defaultPV = await _context.PropertyValues
+                .FirstAsync(x => x.Property.Title == propertyTitle && x.PropertyValue == defaultPropertyValue);
+
+            var productProperties = await _context.ProductProperties
+                .Where(x => x.PropertyValues.PropertyValue == propertyValue)
+                .ToArrayAsync();
+            foreach (var productProperty in productProperties)
+            {
+                productProperty.PropertyValues = defaultPV;
+            }
+
+            await _context.SaveChangesAsync();
+        }
+
         public async Task<bool> AddFilterHeadingOne(string propertyTitle, string headingOneTitle)
         {
             var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
             if (property == null)
                 return false;
-            
+
             var headingOne = await _context.HeadingsOne.FirstOrDefaultAsync(x => x.Title == headingOneTitle);
             if (headingOne == null)
                 return false;
 
-            var headingOneFilter = await _context.HeadingOneFilters.FirstOrDefaultAsync(x => x.heading_one_id == headingOne.Id && x.property_id == property.Id);
+            var headingOneFilter = await _context.HeadingOneFilters.FirstOrDefaultAsync(x =>
+                x.heading_one_id == headingOne.Id && x.property_id == property.Id);
             if (headingOneFilter != null)
                 return true;
-            
+
             headingOneFilter = new HeadingOneFilters() {HeadingOne = headingOne, Property = property};
             await _context.HeadingOneFilters.AddAsync(headingOneFilter);
+
+            var products = await _context.ProductFamilies
+                .Where(x => x.HeadingOne.Title == headingOneTitle)
+                .Join(_context.Products, family => family.Id, product => product.product_family_id,
+                    (family, product) => new {product.Title})
+                .ToListAsync();
+            foreach (var product in products)
+            {
+                await AddProductPropertyValue(product.Title, propertyTitle, "Не указано", false);
+            }
+
             await _context.SaveChangesAsync();
             return true;
         }
-        
+
         public async Task<bool> DeleteHeadingOneFilter(string propertyTitle, string headingOneTitle)
         {
-            var headingOneFilter = await _context.HeadingOneFilters.FirstOrDefaultAsync(x => x.HeadingOne.Title == headingOneTitle && x.Property.Title == propertyTitle);
+            var headingOneFilter = await _context.HeadingOneFilters.FirstOrDefaultAsync(x =>
+                x.HeadingOne.Title == headingOneTitle && x.Property.Title == propertyTitle);
             if (headingOneFilter == null)
                 return true;
 
@@ -214,10 +325,11 @@ namespace BackServer.RepositoryChangers.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
-        
+
         public async Task<bool> DeleteAllHeadingOneFilters(string headingOneTitle)
         {
-            var headingOneFilters = await _context.HeadingOneFilters.Where(x => x.HeadingOne.Title == headingOneTitle).ToArrayAsync();
+            var headingOneFilters = await _context.HeadingOneFilters.Where(x => x.HeadingOne.Title == headingOneTitle)
+                .ToArrayAsync();
             if (headingOneFilters.Length == 0)
                 return true;
 
@@ -225,17 +337,19 @@ namespace BackServer.RepositoryChangers.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
-        
-        public async Task<bool> UpdateHeadingOneFilter(string headingOneFilter, string oldPropertyTitle, string newPropertyTitle)
+
+        public async Task<bool> UpdateHeadingOneFilter(string headingOneFilter, string oldPropertyTitle,
+            string newPropertyTitle)
         {
-            var headingOneFilters = await _context.HeadingOneFilters.FirstOrDefaultAsync(x => x.HeadingOne.Title == headingOneFilter && x.Property.Title == oldPropertyTitle);
+            var headingOneFilters = await _context.HeadingOneFilters.FirstOrDefaultAsync(x =>
+                x.HeadingOne.Title == headingOneFilter && x.Property.Title == oldPropertyTitle);
             if (headingOneFilters == null)
                 return false;
 
             var newProperty = await _context.Properties.FirstOrDefaultAsync(x => x.Title == newPropertyTitle);
             if (newProperty == null)
                 return false;
-            
+
             headingOneFilters.Property = newProperty;
             await _context.SaveChangesAsync();
             return true;
@@ -246,24 +360,37 @@ namespace BackServer.RepositoryChangers.Implementations
             var property = await _context.Properties.FirstOrDefaultAsync(x => x.Title == propertyTitle);
             if (property == null)
                 return false;
-            
+
             var headingTwo = await _context.HeadingsTwo.FirstOrDefaultAsync(x => x.Title == headingTwoTitle);
             if (headingTwo == null)
                 return false;
 
-            var headingOneFilter = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x => x.heading_two_id == headingTwo.Id && x.property_id == property.Id);
+            var headingOneFilter = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x =>
+                x.heading_two_id == headingTwo.Id && x.property_id == property.Id);
             if (headingOneFilter != null)
                 return true;
-            
+
             headingOneFilter = new HeadingTwoFilters() {HeadingTwo = headingTwo, Property = property};
             await _context.HeadingTwoFilters.AddAsync(headingOneFilter);
+
+            var products = await _context.ProductFamilies
+                .Where(x => x.HeadingTwo.Title == headingTwoTitle)
+                .Join(_context.Products, family => family.Id, product => product.product_family_id,
+                    (family, product) => new {product.Title})
+                .ToListAsync();
+            foreach (var product in products)
+            {
+                await AddProductPropertyValue(product.Title, propertyTitle, "Не указано", false);
+            }
+
             await _context.SaveChangesAsync();
             return true;
         }
-        
+
         public async Task<bool> DeleteHeadingTwoFilter(string propertyTitle, string headingTwoTitle)
         {
-            var headingTwoFilters = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x => x.HeadingTwo.Title == headingTwoTitle && x.Property.Title == propertyTitle);
+            var headingTwoFilters = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x =>
+                x.HeadingTwo.Title == headingTwoTitle && x.Property.Title == propertyTitle);
             if (headingTwoFilters == null)
                 return true;
 
@@ -271,10 +398,11 @@ namespace BackServer.RepositoryChangers.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
-        
+
         public async Task<bool> DeleteAllHeadingTwoFilters(string headingTwoTitle)
         {
-            var headingTwoFilters = await _context.HeadingTwoFilters.Where(x => x.HeadingTwo.Title == headingTwoTitle).ToArrayAsync();
+            var headingTwoFilters = await _context.HeadingTwoFilters.Where(x => x.HeadingTwo.Title == headingTwoTitle)
+                .ToArrayAsync();
             if (headingTwoFilters.Length == 0)
                 return true;
 
@@ -282,17 +410,19 @@ namespace BackServer.RepositoryChangers.Implementations
             await _context.SaveChangesAsync();
             return true;
         }
-        
-        public async Task<bool> UpdateHeadingTwoFilter(string headingTwoFilter, string oldPropertyTitle, string newPropertyTitle)
+
+        public async Task<bool> UpdateHeadingTwoFilter(string headingTwoFilter, string oldPropertyTitle,
+            string newPropertyTitle)
         {
-            var headingTwoFilters = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x => x.HeadingTwo.Title == headingTwoFilter && x.Property.Title == oldPropertyTitle);
+            var headingTwoFilters = await _context.HeadingTwoFilters.FirstOrDefaultAsync(x =>
+                x.HeadingTwo.Title == headingTwoFilter && x.Property.Title == oldPropertyTitle);
             if (headingTwoFilters == null)
                 return false;
 
             var newProperty = await _context.Properties.FirstOrDefaultAsync(x => x.Title == newPropertyTitle);
             if (newProperty == null)
                 return false;
-            
+
             headingTwoFilters.Property = newProperty;
             await _context.SaveChangesAsync();
             return true;
